+++ src/physfs/physfs_sdl.cpp
@@ -58,7 +58,7 @@ static int func_read(struct SDL_RWops* context, void* ptr, int size, int maxnum)
 {
     PHYSFS_file* file = (PHYSFS_file*) context->hidden.unknown.data1;
 
-    int res = PHYSFS_read(file, ptr, size, maxnum);
+    int res = (int)PHYSFS_read(file, ptr, size, maxnum);
     return res;
 }
