+++ src/worldobjs/conveyor_belt.cpp
@@ -43,12 +43,12 @@ ConveyorBelt::draw (SceneContext& gc)
   gc.color().draw(left_sur, pos);
   for (int i=0; i < width; ++i)
     gc.color().draw(middle_sur,
-	    Vector3f(static_cast<int>(pos.x + left_sur.get_width() + i * middle_sur.get_width()),
-                  static_cast<int>(pos.y)));
+	    Vector3f(static_cast<float>(pos.x + left_sur.get_width() + i * middle_sur.get_width()),
+                  static_cast<float>(pos.y)));
 
   gc.color().draw(right_sur,
-	  Vector3f(static_cast<int>(pos.x + left_sur.get_width() + width * middle_sur.get_width()),
-                 static_cast<int>(pos.y)));
+	  Vector3f(static_cast<float>(pos.x + left_sur.get_width() + width * middle_sur.get_width()),
+                 static_cast<float>(pos.y)));
 }
 
 void
